refactor: 파일이름 변경
- constants/httpError.ts 를 constants/error.ts 로 변경
